+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @vladmandic/human  
 
-  Version: **3.3.4**  
+  Version: **3.3.5**  
   Description: **Human: AI-powered 3D Face Detection & Rotation Tracking, Face Description & Recognition, Body Pose Tracking, 3D Hand & Finger Tracking, Iris Analysis, Age & Gender & Emotion Prediction, Gesture Recognition**  
   
   Author: **Vladimir Mandic <mandic00@live.com>**  
@@ -9,7 +9,10 @@
   
 ## Changelog
   
-### **HEAD -> main** 2024/10/24 mandic00@live.com
+### **3.3.5** 2025/02/05 mandic00@live.com
+
+
+### **origin/main** 2024/10/24 mandic00@live.com
 
 - add human.draw.tensor method